finished

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-         $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
-
+            $data = [
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
     }
+}
+    
 
     /**
      * Show the form for creating a new resource.
@@ -36,6 +43,8 @@ class TasksController extends Controller
             'task' => $task,
         ]);
 
+     return redirect()->back();
+        
     }
 
     /**
@@ -52,12 +61,14 @@ class TasksController extends Controller
         ]);
 
         
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
 
-        return redirect('/');
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            
+        ]);
+
+        return redirect()->back();
 
     }
 
@@ -69,14 +80,19 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        
+       $task = \App\Task::find($id);
+       if (\Auth::id() != $task->user_id) {
+              return redirect('/');
+        }
+       else{ 
+        
         $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
-
-    }
-
+        $data = [
+                 'task' => $task,
+        ];
+       return view('tasks.show', $data);
+    }}
     /**
      * Show the form for editing the specified resource.
      *
@@ -126,9 +142,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
 
-        return redirect('/');
+         if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
 
     }
 }
